Cppcheck & Doxygen update: * Comments refactoring * arm_macros.h: BYTE_TO_PERC test >= 255 which doesn't raise warning when given BYTE parameter * arm_macros.h: re-enabled TEST_BITS_SET & TEST_BITS_CLR macros (TEST_BITS_VAL set to be alias of TEST_BITS_SET) * Few MISRA compliance changes

+++ arm_inlines_angle.h
@@ -37,7 +37,7 @@
 **/
 __INLINE float RAD_NORM(const float rad)
 {
-	const float val = (rad < 0.0f) ? M_TWOPI : -M_TWOPI;
+	const float val = (float) ((rad < 0.0f) ? M_TWOPI : -M_TWOPI);
 	float r = rad;
 
 	while (fabs(r) > M_TWOPI)	{ r += val; }
